+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.20.0] - 2026-08-27
+## [0.20.2] - 2026-08-31
+
+### Added
+
+- `FilterMarkers` to keep markers whose CPM exceeds an isotype-relative
+  ratio and/or an absolute abundance threshold in at least
+  `min_cell_fraction` of cells.
+
+## [0.20.1] - 2026-08-27
 
 ### Added